breweries_project/requirements.txt
pandas
requests
pyarrow<10.1.0,>=10.0.1
apache-airflow